+++ package.json
@@ -9,10 +9,11 @@
   "author": "shivanand",
   "license": "ISC",
   "dependencies": {
-    "convict": "^5.1.0",
+    "convict": "^6.0.0",
     "dotenv": "^8.1.0",
     "express": "^4.17.1",
     "mongodb": "^3.5.9",
+    "mongoose": "^5.9.26",
     "mysql": "^2.18.1"
   }
 }
